+++ components/bridge/connection/sample/shared/src/androidMain/kotlin/com/flipperdevices/bridge/connection/screens/search/BLESearchConnectionDelegate.kt
@@ -21,7 +21,6 @@ import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onEach
-import me.gulya.anvil.assisted.ContributesAssistedFactory
 import no.nordicsemi.android.kotlin.ble.core.ServerDevice
 import no.nordicsemi.android.kotlin.ble.core.scanner.BleScanMode
 import no.nordicsemi.android.kotlin.ble.core.scanner.BleScannerSettings