+++ components/bridge/connection/transport/usb/impl/src/androidMain/kotlin/com/flipperdevices/bridge/connection/transport/usb/impl/model/USBAndroidDeviceFactory.kt
@@ -9,7 +9,6 @@ import com.squareup.anvil.annotations.ContributesBinding
 import kotlinx.coroutines.CoroutineScope
 import javax.inject.Inject
 
-
 @ContributesBinding(AppGraph::class, USBPlatformDeviceFactory::class)
 class USBAndroidDeviceFactory @Inject constructor(
     private val context: Context