+++ components/bridge/connection/transport/usb/impl/src/jvmSharedMain/kotlin/com/flipperdevices/bridge/connection/transport/usb/impl/di/BleDeviceConnectionModule.kt
@@ -5,7 +5,6 @@ import com.flipperdevices.bridge.connection.transport.common.api.di.DeviceConnec
 import com.flipperdevices.bridge.connection.transport.common.api.di.toHolder
 import com.flipperdevices.bridge.connection.transport.usb.api.FUSBDeviceConnectionConfig
 import com.flipperdevices.bridge.connection.transport.usb.impl.USBDeviceConnectionApiImpl
-import com.flipperdevices.bridge.connection.transport.usb.impl.model.USBPlatformDevice
 import com.flipperdevices.bridge.connection.transport.usb.impl.model.USBPlatformDeviceFactory
 import com.flipperdevices.core.di.AppGraph
 import com.squareup.anvil.annotations.ContributesTo